refactor(styles): address naming conventions to style files #52
There are a lot of files in the ./src directory, this commit updates
the names of those files to make it more clear as to what files they are
and what they are used for.

style.scss was renamed to style-base.scss as a simple fix to a
postCss problem.

Changes to be committed:
modified:   docs/api.md
modified:   src/auro-menu-option.js
modified:   src/auro-menu.js
modified:   src/auro-sub-menu.js
renamed:    src/style-fixed.scss -> src/style-base-fixed.scss
renamed:    src/style.scss -> src/style-base.scss
renamed:    src/auro-menu-option-fixed.scss -> src/style-menu-option-fixed.scss
renamed:    src/auro-menu-option.scss -> src/style-menu-option.scss
renamed:    src/auro-sub-menu-fixed.scss -> src/style-sub-menu-fixed.scss
renamed:    src/auro-sub-menu.scss -> src/style-sub-menu.scss
