+++ src/style-base-fixed.scss
@@ -1,5 +1,5 @@
 @import './../node_modules/@alaskaairux/design-tokens/dist/tokens/SassCustomProperties';
 
 :host([fixed]) {
-  @import './style.scss';
+  @import './style-base.scss';
 }
