+++ src/style-menu-option-fixed.scss
@@ -1,5 +1,5 @@
 @import './../node_modules/@alaskaairux/design-tokens/dist/tokens/SassCustomProperties';
 
 :host([fixed]) {
-  @import './auro-sub-menu.scss';
+  @import './style-menu-option.scss';
 }
