+++ src/style-sub-menu-fixed.scss
@@ -1,5 +1,5 @@
 @import './../node_modules/@alaskaairux/design-tokens/dist/tokens/SassCustomProperties';
 
 :host([fixed]) {
-  @import './auro-menu-option.scss';
+  @import './style-sub-menu.scss';
 }
